anika-copy/requirements.txt
streamlit
numpy
pandas
matplotlib
seaborn
scipy
scikit-learn
plotly.express
imbalanced-learn